Move Now Playing to a bottom sheet fragment
