+++ project/app/src/main/java/com/iven/musicplayergo/interfaces/MediaPlayerInterface.kt
@@ -8,6 +8,9 @@ interface MediaPlayerInterface {
     fun onStartPlayback(song: Music?, songs: List<Music>?, launchedBy: LaunchedBy)
     fun onStateChanged()
     fun onPositionChanged(position: Int)
+    fun onSetRepeatNP()
+    fun onOpenPlayingArtistAlbumNP()
+    fun onSkipNP(isNext: Boolean)
     fun onPlaybackCompleted()
     fun onUpdateRepeatStatus()
     fun onShuffleSongs(songs: MutableList<Music>?, launchedBy: LaunchedBy)
