+++ project/app/src/main/res/layout-land/now_playing.xml
@@ -1,10 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    android:clickable="false"
-    android:focusable="false"
     android:fitsSystemWindows="true"
     android:orientation="vertical">
 
@@ -19,11 +16,13 @@
             android:layout_height="wrap_content"
             android:layout_marginBottom="8dp"
             android:layout_weight="1"
+            android:id="@+id/np_playing_info"
             android:background="@drawable/ripple"
             android:orientation="vertical"
-            android:paddingTop="8dp"
-            android:paddingBottom="8dp"
-            tools:context=".ui.MainActivity">
+            android:paddingTop="4dp"
+            android:paddingBottom="4dp"
+            android:clickable="true"
+            android:focusable="true">
 
             <TextView
                 android:id="@+id/np_song"
