+++ project/app/src/main/res/values-night/styles.xml
@@ -1,0 +1,9 @@
+<resources>
+
+    <!-- set the rounded drawable as background to your bottom sheet -->
+    <style name="BaseBottomSheetDialog" parent="@style/Theme.MaterialComponents.BottomSheetDialog">
+        <item name="android:windowIsFloating">false</item>
+        <item name="bottomSheetStyle">@style/BottomSheet</item>
+        <item name="colorAccent">?attr/colorPrimary</item>
+    </style>
+</resources>
